Move core client to its own context
