+++ ui/App.tsx
@@ -15,6 +15,7 @@ import ErrorBoundary from "./components/ErrorBoundary";
 import Layout from "./components/Layout";
 import AppContextProvider from "./contexts/AppContext";
 import AuthContextProvider, { AuthCheck } from "./contexts/AuthContext";
+import CoreClientContextProvider from "./contexts/CoreClientContext";
 import FeatureFlagsContextProvider from "./contexts/FeatureFlags";
 import { Core } from "./lib/api/core/core.pb";
 import Fonts from "./lib/fonts";
@@ -103,19 +104,21 @@ export default function AppContainer() {
           <Fonts />
           <GlobalStyle />
           <Router>
-            <AppContextProvider renderFooter coreClient={Core}>
+            <AppContextProvider renderFooter>
               <FeatureFlagsContextProvider>
                 <AuthContextProvider>
-                  <Switch>
-                    {/* <Signin> does not use the base page <Layout> so pull it up here */}
-                    <Route component={SignIn} exact path="/sign_in" />
-                    <Route path="*">
-                      {/* Check we've got a logged in user otherwise redirect back to signin */}
-                      <AuthCheck>
-                        <App />
-                      </AuthCheck>
-                    </Route>
-                  </Switch>
+                  <CoreClientContextProvider api={Core}>
+                    <Switch>
+                      {/* <Signin> does not use the base page <Layout> so pull it up here */}
+                      <Route component={SignIn} exact path="/sign_in" />
+                      <Route path="*">
+                        {/* Check we've got a logged in user otherwise redirect back to signin */}
+                        <AuthCheck>
+                          <App />
+                        </AuthCheck>
+                      </Route>
+                    </Switch>
+                  </CoreClientContextProvider>
                 </AuthContextProvider>
               </FeatureFlagsContextProvider>
             </AppContextProvider>
