+++ ui/contexts/AppContext.tsx
@@ -1,8 +1,6 @@
-import _ from "lodash";
 import * as React from "react";
 import { useHistory } from "react-router-dom";
 import { Applications } from "../lib/api/applications/applications.pb";
-import { Core } from "../lib/api/core/core.pb";
 import { formatURL } from "../lib/nav";
 import {
   clearCallbackState,
@@ -13,7 +11,6 @@ import {
 } from "../lib/storage";
 import { PageRoute, V2Routes } from "../lib/types";
 import { notifySuccess } from "../lib/utils";
-import { AuthRoutes } from "./AuthContext";
 
 type AppState = {
   error: null | { fatal: boolean; message: string; detail?: string };
@@ -31,7 +28,6 @@ export function defaultLinkResolver(incoming: string): string {
 
 export type AppContextType = {
   applicationsClient: typeof Applications;
-  api: typeof Core;
   userConfigRepoName: string;
   doAsyncError: (message: string, detail: string) => void;
   clearAsyncError: () => void;
@@ -57,7 +53,6 @@ export const AppContext = React.createContext<AppContextType>(
 
 export interface AppProps {
   applicationsClient?: typeof Applications;
-  coreClient?: typeof Core;
   linkResolver?: LinkResolver;
   children?: any;
   renderFooter?: boolean;
@@ -93,23 +88,8 @@ export default function AppContextProvider({
     });
   };
 
-  const wrapped = {} as typeof Core;
-
-  //   Wrap each API method in a check that redirects to the signin page if a 401 is returned.
-  _.each(props.coreClient, (fn: any, method) => {
-    wrapped[method] = (req, initReq) => {
-      return fn(req, initReq).catch((err) => {
-        if (err.code === 401) {
-          history.push(AuthRoutes.AUTH_PATH_SIGNIN);
-        }
-        throw err;
-      });
-    };
-  });
-
   const value: AppContextType = {
     applicationsClient,
-    api: wrapped,
     userConfigRepoName: "wego-github-jlw-config-repo",
     doAsyncError,
     clearAsyncError,
